% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.url.R
\name{base.url}
\alias{base.url}
\title{Base URL Construction Method}
\usage{
base.url(end.point)
}
\arguments{
\item{end.point}{The url that should be used in contact the Thor server. 
When Thor is deployed locally, it make sense to use the localhost 
designation. On the other hand, when deployed remotely, a particular URL 
needs to be provided.}
}
\value{
A string containing the URL endpoint of the desired API method.
}
\description{
Base URL Construction Method
}
