% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thor.client.R
\name{create.experiment}
\alias{create.experiment}
\title{Thor Client Methods}
\usage{
create.experiment(auth.token, name, dimensions, acq.func, overwrite)
}
\arguments{
\item{auth.token}{String containing a user's specific API key provided by 
the Thor server. This is used to authenticate with the Thor server as a 
handshake that these experiments belong to a user and can be viewed and 
edited by them.}

\item{name}{String containing the name of the experiment to create.}

\item{dimensions}{A list of dictionaries that specify the hyperparameters
of the machine learning system.}

\item{acq.func}{A string containing the name of the acquisition function 
to use. This can be one of "hedge", "upper_confidence_bound", 
"expected_improvement", or "improvement_probability".}

\item{overwrite}{An indicator variable which will overwrite existing 
experiments with the given name if they already exist on Thor Server.}
}
\description{
The purpose of this module is to allow Python users to utilize the Thor 
utility for Bayesian optimization of machine learning systems. This module
in particular is intended to properly authenticate with Thor and to create
experiments.
Create an experiment.
}
