% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.client.R
\name{create.recommendation}
\alias{create.recommendation}
\title{Thor Experiment Client Methods}
\usage{
create.recommendation(experiment.id, auth.token, rand.prob, n.model.iters)
}
\arguments{
\item{experiment.id}{A unique identifier that indicates which experiment
on the server-side is being interacted with by the client.}

\item{auth.token}{String containing a user's specific API key provided by 
the Thor server. This is used to authenticate with the Thor server as a 
handshake that these experiments belong to a user and can be viewed and 
edited by them.}

\item{rand.prob}{This parameter represents that a random point in the input
space is chosen instead of selecting a configuration of parameters using 
Bayesian optimization. As such, this parameter can be used to benchmark 
against random search and otherwise to perform pure exploration of the
parameter space.}

\item{n.model.iters}{This parameter determines the number of maximum 
likelihood random restarts are used to isolate the maximum of the Gaussian
process likelihood with respect to the kernel parameters. Setting this to 
a large value will generally provide better probabilistic interpolants of 
the metric as a function of the model parameters. In large-scale problems,
this parameter instead controls the number of training iterations used to 
fit a Bayesian neural network. In particular, 1,000 times this parameter 
epochs are performed.}
}
\value{
A recommendation client object corresponding to the recommended 
set of parameters.
}
\description{
This object defines a Thor experiment within the Python environment. In
particular, an experiment is defined by its name, the date at which it was
created, and the dimensions of the machine learning model. Moreover, an
authentication token is required for requesting new parameter
configurations, for submitting observations of parameters, for viewing
pending parameter configurations and for obtaining the best configuration
of parameters that has been evaluated so far.
Get a recommendation for a point to evaluate next.
}
\details{
The create recommendation utility represents the core of the Thor Bayesian
optimization software. This function will contact the Thor server and 
request a new configuration of machine learning parameters that serve the 
object of maximizing the metric of interest.
}
