% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thor.client.R
\name{experiment.for.name}
\alias{experiment.for.name}
\title{Get an experiment with a given name.}
\usage{
experiment.for.name(auth.token, name)
}
\arguments{
\item{auth.token}{String containing a user's specific API key provided by 
the Thor server. This is used to authenticate with the Thor server as a 
handshake that these experiments belong to a user and can be viewed and 
edited by them.}

\item{name}{String containing the name of the experiment that should be 
obtained.}
}
\value{
corresponding experiment with the provided name. The object returned 
by this method is identical in functionality to the object that is
returned by the function `create_experiment`.
}
\description{
Get an experiment with a given name.
}
