% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.client.R
\name{pending.recommendations}
\alias{pending.recommendations}
\title{Query for pending recommendations that have yet to be evaluated.
Upload a pairing of a configuration alongside an observed target variable.}
\usage{
pending.recommendations(experiment.id, auth.token)
}
\arguments{
\item{experiment.id}{A unique identifier that indicates which experiment
on the server-side is being interacted with by the client.}

\item{auth.token}{String containing a user's specific API key provided by 
the Thor server. This is used to authenticate with the Thor server as a 
handshake that these experiments belong to a user and can be viewed and 
edited by them.}
}
\value{
A list of recommendation client objects, where each element in the
list corresponds to a pending observation.
}
\description{
Sometimes client-side computations may fail for a given input
configuration of model parameters, leaving the recommendation in a kind
of "limbo" state in which is not being evaluated but still exists. In
this case, it can be advantageous for the client to query for such
pending observations and to evaluate them. This function returns a list
of pending recommendations which can then be evaluated by the client.
}
