% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.client.R
\name{submit.observation}
\alias{submit.observation}
\title{Upload a pairing of a configuration alongside an observed target variable.}
\usage{
submit.observation(experiment.id, auth.token, config, target)
}
\arguments{
\item{experiment.id}{A unique identifier that indicates which experiment
on the server-side is being interacted with by the client.}

\item{auth.token}{String containing a user's specific API key provided by 
the Thor server. This is used to authenticate with the Thor server as a 
handshake that these experiments belong to a user and can be viewed and 
edited by them.}

\item{config}{A dictionary mapping dimension names to values indicating 
the configuration of parameters.}

\item{target}{A number indicating the performance of this configuration 
of model parameters.}
}
\description{
Upload a pairing of a configuration alongside an observed target variable.
}
