% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommendation.client.R
\name{submit.recommendation}
\alias{submit.recommendation}
\title{Thor Recommendation Client Methods}
\usage{
submit.recommendation(recommendation.id, auth.token, value)
}
\arguments{
\item{recommendation.id}{An integer identifier for this recommendation. 
This allows the Thor server to identify which recommendation was 
associated with a given value of the metric on the client side.}

\item{auth.token}{String containing a user's specific API key provided by 
the Thor server. This is used to authenticate with the Thor server as a 
handshake that these experiments belong to a user and can be viewed and 
edited by them.}

\item{value}{A number indicating the performance of this configuration of 
model parameters.}
}
\value{
A dictionary containing the recommendation identifier and a 
boolean indicator that the recommendation was submitted.
}
\description{
The purpose of this class is to provide an interface for interacting with 
suggestions provided by the Thor API. In particular, once a recommendation
is received, it will be evaluated by the client's local computer. After 
this, its value as a configuration will be transmitted back to the Thor 
server.
Submit the returned metric value for a point that was recommended by the 
Bayesian optimization routine.
}
